requirements.txt
asgiref==3.7.2
Django==3.2.23
django-cors-headers==4.3.0
djangorestframework==3.14.0
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
